refactor: cleanup api module structure
